RBM as interface
