+++ hadoop-tools/hadoop-azure/src/main/java/org/apache/hadoop/fs/azurebfs/services/ReadBufferWorker.java
@@ -53,7 +53,7 @@ class ReadBufferWorker implements Runnable {
     } catch (InterruptedException ex) {
       Thread.currentThread().interrupt();
     }
-    ReadBufferManager bufferManager = ReadBufferManager.getBufferManager();
+    ReadBufferManagerV1 bufferManager = ReadBufferManagerV1.getBufferManager();
     ReadBuffer buffer;
     while (isRunning.get()) {
       try {
@@ -75,7 +75,7 @@ class ReadBufferWorker implements Runnable {
               Math.min(buffer.getRequestedLength(), buffer.getBuffer().length),
                   buffer.getTracingContext());
 
-          bufferManager.doneReading(buffer, ReadBufferStatus.AVAILABLE, bytesRead);  // post result back to ReadBufferManager
+          bufferManager.doneReading(buffer, ReadBufferStatus.AVAILABLE, bytesRead);  // post result back to ReadBufferManagerV1
         } catch (IOException ex) {
           buffer.setErrException(ex);
           bufferManager.doneReading(buffer, ReadBufferStatus.READ_FAILED, 0);
