+++ hadoop-tools/hadoop-azure/src/test/java/org/apache/hadoop/fs/azurebfs/services/TestAbfsInputStream.java
@@ -89,7 +89,7 @@ public class TestAbfsInputStream extends
   @Override
   public void teardown() throws Exception {
     super.teardown();
-    ReadBufferManager.getBufferManager().testResetReadBufferManager();
+    ReadBufferManagerV1.getBufferManager().testResetReadBufferManager();
   }
 
   private AbfsRestOperation getMockRestOp() {
@@ -164,12 +164,12 @@ public class TestAbfsInputStream extends
 
   private void queueReadAheads(AbfsInputStream inputStream) {
     // Mimic AbfsInputStream readAhead queue requests
-    ReadBufferManager.getBufferManager()
+    ReadBufferManagerV1.getBufferManager()
         .queueReadAhead(inputStream, 0, ONE_KB, inputStream.getTracingContext());
-    ReadBufferManager.getBufferManager()
+    ReadBufferManagerV1.getBufferManager()
         .queueReadAhead(inputStream, ONE_KB, ONE_KB,
             inputStream.getTracingContext());
-    ReadBufferManager.getBufferManager()
+    ReadBufferManagerV1.getBufferManager()
         .queueReadAhead(inputStream, TWO_KB, TWO_KB,
             inputStream.getTracingContext());
   }
@@ -187,15 +187,16 @@ public class TestAbfsInputStream extends
   private void checkEvictedStatus(AbfsInputStream inputStream, int position, boolean expectedToThrowException)
       throws Exception {
     // Sleep for the eviction threshold time
-    Thread.sleep(ReadBufferManager.getBufferManager().getThresholdAgeMilliseconds() + 1000);
+    Thread.sleep(
+        ReadBufferManagerV1.getBufferManager().getThresholdAgeMilliseconds() + 1000);
 
     // Eviction is done only when AbfsInputStream tries to queue new items.
     // 1 tryEvict will remove 1 eligible item. To ensure that the current test buffer
     // will get evicted (considering there could be other tests running in parallel),
     // call tryEvict for the number of items that are there in completedReadList.
-    int numOfCompletedReadListItems = ReadBufferManager.getBufferManager().getCompletedReadListSize();
+    int numOfCompletedReadListItems = ReadBufferManagerV1.getBufferManager().getCompletedReadListSize();
     while (numOfCompletedReadListItems > 0) {
-      ReadBufferManager.getBufferManager().callTryEvict();
+      ReadBufferManagerV1.getBufferManager().callTryEvict();
       numOfCompletedReadListItems--;
     }
 
@@ -210,7 +211,7 @@ public class TestAbfsInputStream extends
   public TestAbfsInputStream() throws Exception {
     super();
     // Reduce thresholdAgeMilliseconds to 3 sec for the tests
-    ReadBufferManager.getBufferManager().setThresholdAgeMilliseconds(REDUCED_READ_BUFFER_AGE_THRESHOLD);
+    ReadBufferManagerV1.getBufferManager().setThresholdAgeMilliseconds(REDUCED_READ_BUFFER_AGE_THRESHOLD);
   }
 
   private void writeBufferToNewFile(Path testFile, byte[] buffer) throws IOException {
@@ -319,7 +320,7 @@ public class TestAbfsInputStream extends
    * This test expects AbfsInputStream to throw the exception that readAhead
    * thread received on read. The readAhead thread must be initiated from the
    * active read request itself.
-   * Also checks that the ReadBuffers are evicted as per the ReadBufferManager
+   * Also checks that the ReadBuffers are evicted as per the ReadBufferManagerV1
    * threshold criteria.
    * @throws Exception
    */
@@ -364,7 +365,7 @@ public class TestAbfsInputStream extends
   public void testFailedReadAheadEviction() throws Exception {
     AbfsClient client = getMockAbfsClient();
     AbfsRestOperation successOp = getMockRestOp();
-    ReadBufferManager.setThresholdAgeMilliseconds(INCREASED_READ_BUFFER_AGE_THRESHOLD);
+    ReadBufferManagerV1.setThresholdAgeMilliseconds(INCREASED_READ_BUFFER_AGE_THRESHOLD);
     // Stub :
     // Read request leads to 3 readahead calls: Fail all 3 readahead-client.read()
     // Actual read request fails with the failure in readahead thread
@@ -379,15 +380,15 @@ public class TestAbfsInputStream extends
     // Add a failed buffer to completed queue and set to no free buffers to read ahead.
     ReadBuffer buff = new ReadBuffer();
     buff.setStatus(ReadBufferStatus.READ_FAILED);
-    ReadBufferManager.getBufferManager().testMimicFullUseAndAddFailedBuffer(buff);
+    ReadBufferManagerV1.getBufferManager().testMimicFullUseAndAddFailedBuffer(buff);
 
     // if read failed buffer eviction is tagged as a valid eviction, it will lead to
     // wrong assumption of queue logic that a buffer is freed up and can lead to :
     // java.util.EmptyStackException
     // at java.util.Stack.peek(Stack.java:102)
     // at java.util.Stack.pop(Stack.java:84)
-    // at org.apache.hadoop.fs.azurebfs.services.ReadBufferManager.queueReadAhead
-    ReadBufferManager.getBufferManager().queueReadAhead(inputStream, 0, ONE_KB,
+    // at org.apache.hadoop.fs.azurebfs.services.ReadBufferManagerV1.queueReadAhead
+    ReadBufferManagerV1.getBufferManager().queueReadAhead(inputStream, 0, ONE_KB,
         getTestTracingContext(getFileSystem(), true));
   }
 
@@ -395,7 +396,7 @@ public class TestAbfsInputStream extends
    *
    * The test expects AbfsInputStream to initiate a remote read request for
    * the request offset and length when previous read ahead on the offset had failed.
-   * Also checks that the ReadBuffers are evicted as per the ReadBufferManager
+   * Also checks that the ReadBuffers are evicted as per the ReadBufferManagerV1
    * threshold criteria.
    * @throws Exception
    */
@@ -429,7 +430,7 @@ public class TestAbfsInputStream extends
     verifyReadCallCount(client, 3);
 
     // Sleep for thresholdAgeMs so that the read ahead buffer qualifies for being old.
-    Thread.sleep(ReadBufferManager.getBufferManager().getThresholdAgeMilliseconds());
+    Thread.sleep(ReadBufferManagerV1.getBufferManager().getThresholdAgeMilliseconds());
 
     // Second read request should retry the read (and not issue any new readaheads)
     inputStream.read(ONE_KB, new byte[ONE_KB], 0, ONE_KB);
@@ -474,7 +475,7 @@ public class TestAbfsInputStream extends
             any(String.class), any(), any(TracingContext.class));
 
     AbfsInputStream inputStream = getAbfsInputStream(client, "testSuccessfulReadAhead.txt");
-    int beforeReadCompletedListSize = ReadBufferManager.getBufferManager().getCompletedReadListSize();
+    int beforeReadCompletedListSize = ReadBufferManagerV1.getBufferManager().getCompletedReadListSize();
 
     // First read request that triggers readAheads.
     inputStream.read(new byte[ONE_KB]);
@@ -482,9 +483,9 @@ public class TestAbfsInputStream extends
     // Only the 3 readAhead threads should have triggered client.read
     verifyReadCallCount(client, 3);
     int newAdditionsToCompletedRead =
-        ReadBufferManager.getBufferManager().getCompletedReadListSize()
+        ReadBufferManagerV1.getBufferManager().getCompletedReadListSize()
             - beforeReadCompletedListSize;
-    // read buffer might be dumped if the ReadBufferManager getblock preceded
+    // read buffer might be dumped if the ReadBufferManagerV1 getblock preceded
     // the action of buffer being picked for reading from readaheadqueue, so that
     // inputstream can proceed with read and not be blocked on readahead thread
     // availability. So the count of buffers in completedReadQueue for the stream
@@ -529,8 +530,8 @@ public class TestAbfsInputStream extends
             any(String.class), nullable(ContextEncryptionAdapter.class),
             any(TracingContext.class));
 
-    final ReadBufferManager readBufferManager
-        = ReadBufferManager.getBufferManager();
+    final ReadBufferManagerV1 readBufferManager
+        = ReadBufferManagerV1.getBufferManager();
 
     final int readBufferTotal = readBufferManager.getNumBuffers();
     final int expectedFreeListBufferCount = readBufferTotal
@@ -574,7 +575,7 @@ public class TestAbfsInputStream extends
   /**
    * This test expects ReadAheadManager to throw exception if the read ahead
    * thread had failed within the last thresholdAgeMilliseconds.
-   * Also checks that the ReadBuffers are evicted as per the ReadBufferManager
+   * Also checks that the ReadBuffers are evicted as per the ReadBufferManagerV1
    * threshold criteria.
    * @throws Exception
    */
@@ -607,7 +608,7 @@ public class TestAbfsInputStream extends
     // if readAhead failed for specific offset, getBlock should
     // throw exception from the ReadBuffer that failed within last thresholdAgeMilliseconds sec
     intercept(IOException.class,
-        () -> ReadBufferManager.getBufferManager().getBlock(
+        () -> ReadBufferManagerV1.getBufferManager().getBlock(
             inputStream,
             0,
             ONE_KB,
@@ -624,7 +625,7 @@ public class TestAbfsInputStream extends
   /**
    * The test expects ReadAheadManager to return 0 receivedBytes when previous
    * read ahead on the offset had failed and not throw exception received then.
-   * Also checks that the ReadBuffers are evicted as per the ReadBufferManager
+   * Also checks that the ReadBuffers are evicted as per the ReadBufferManagerV1
    * threshold criteria.
    * @throws Exception
    */
@@ -655,14 +656,14 @@ public class TestAbfsInputStream extends
     // AbfsInputStream Read would have waited for the read-ahead for the requested offset
     // as we are testing from ReadAheadManager directly, sleep for thresholdAgeMilliseconds so that
     // read buffer qualifies for to be an old buffer
-    Thread.sleep(ReadBufferManager.getBufferManager().getThresholdAgeMilliseconds());
+    Thread.sleep(ReadBufferManagerV1.getBufferManager().getThresholdAgeMilliseconds());
 
     // Only the 3 readAhead threads should have triggered client.read
     verifyReadCallCount(client, 3);
 
     // getBlock from a new read request should return 0 if there is a failure
     // 30 sec before in read ahead buffer for respective offset.
-    int bytesRead = ReadBufferManager.getBufferManager().getBlock(
+    int bytesRead = ReadBufferManagerV1.getBufferManager().getBlock(
         inputStream,
         ONE_KB,
         ONE_KB,
@@ -715,7 +716,7 @@ public class TestAbfsInputStream extends
     verifyReadCallCount(client, 3);
 
     // getBlock for a new read should return the buffer read-ahead
-    int bytesRead = ReadBufferManager.getBufferManager().getBlock(
+    int bytesRead = ReadBufferManagerV1.getBufferManager().getBlock(
         inputStream,
         ONE_KB,
         ONE_KB,
@@ -853,7 +854,7 @@ public class TestAbfsInputStream extends
         .describedAs("Unexpected AlwaysReadBufferSize settings")
         .isEqualTo(alwaysReadBufferSizeEnabled);
 
-    Assertions.assertThat(ReadBufferManager.getBufferManager().getReadAheadBlockSize())
+    Assertions.assertThat(ReadBufferManagerV1.getBufferManager().getReadAheadBlockSize())
         .describedAs("Unexpected readAhead block size")
         .isEqualTo(readAheadBlockSize);
 
@@ -921,9 +922,9 @@ public class TestAbfsInputStream extends
   }
 
   private void resetReadBufferManager(int bufferSize, int threshold) {
-    ReadBufferManager.getBufferManager()
+    ReadBufferManagerV1.getBufferManager()
         .testResetReadBufferManager(bufferSize, threshold);
-    // Trigger GC as aggressive recreation of ReadBufferManager buffers
+    // Trigger GC as aggressive recreation of ReadBufferManagerV1 buffers
     // by successive tests can lead to OOM based on the dev VM/machine capacity.
     System.gc();
   }
